.history/requirements_20250709112423.txt
streamlit
pillow
numpy
google-generativeai
python-dotenv